+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.90.0
+
 - Skills support following https://agentskills.io . #241
   - Make skills available via commands as well e.g. `/my-skill`.
   - new command `/skills` to list available skills.